Revert "cleanup: change Set/GetIntrinsic in ArtMethod to use Int..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25464199-7def-42ba-bad3-6a663e9aa46a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

+++ runtime/runtime_intrinsics.cc
@@ -63,10 +63,10 @@ bool InitializeIntrinsic(Thread* self,
 
   CHECK_EQ(method->GetInvokeType(), invoke_type);
   if (method->IsIntrinsic()) {
-    CHECK_EQ(method->GetIntrinsic(), intrinsic);
+    CHECK_EQ(method->GetIntrinsic(), static_cast<uint32_t>(intrinsic));
     return true;
   } else {
-    method->SetIntrinsic(intrinsic);
+    method->SetIntrinsic(static_cast<uint32_t>(intrinsic));
     return false;
   }
 }
@@ -83,7 +83,7 @@ bool IsIntrinsicInitialized(Thread* self,
 
   CHECK_EQ(method->GetInvokeType(), invoke_type);
   if (method->IsIntrinsic()) {
-    CHECK_EQ(method->GetIntrinsic(), intrinsic);
+    CHECK_EQ(method->GetIntrinsic(), static_cast<uint32_t>(intrinsic));
     return true;
   } else {
     return false;
